refactor MeleeChemicalInjector
